+++ tests/test_130-profiles.py
@@ -36,13 +36,21 @@ def test_update(cached_profile):
 def test_set_scopes(cached_profile, cached_environment):
     scopes = britive.profiles.set_scopes(
         profile_id=cached_profile['papId'],
-        scopes= [
+        scopes=[
             {
                 'type': 'Environment',
                 'value': cached_environment['id']
             }
         ]
     )
+    assert isinstance(scopes, list)
+    assert len(scopes) == 1
+    assert isinstance(scopes[0], dict)
+    assert scopes[0]['value'] == cached_environment['id']
+
+
+def test_get_scopes(cached_profile, cached_environment):
+    scopes = britive.profiles.get_scopes(profile_id=cached_profile['papId'])
     assert isinstance(scopes, list)
     assert len(scopes) == 1
     assert isinstance(scopes[0], dict)