+++ CHANGELOG.md
@@ -2,17 +2,31 @@
 
 All changes to the package starting with v2.8.1 will be logged here.
 
-## v2.6.0 [2023-02-06]
+## v2.16.0 [2023-03-02]
 #### What's New
-* Added APIs for workload federation administrative functions
-  * `profiles.add_single_environment_scope()`
-  * `profiles.remove_single_environment_scope()`
+* Natively support Azure Managed Identity OIDC authentication for workload federation.
 
 #### Enhancements
 * None
 
 #### Bug Fixes
 * None
+
+#### Dependencies
+* None
+
+#### Other
+* None
+
+## v2.15.1 [2023-02-16]
+#### What's New
+* None
+
+#### Enhancements
+* None
+
+#### Bug Fixes
+* Added missing API call `profiles.get_scopes()`
 
 #### Dependencies
 * None